fix: LSDV-4600: Lead time for TextArea (#1214)
Annotation results now have a "lead_time"value in export data for textareas. The value reflects how much time the annotator spent typing values in this TextArea. Lead Time is calculated with minimal interaction time 0.5s and with debounce of the same 0.5s, for example:
- slow typer, one letter per second, 10 letters: 10*0.5s = 5s lead time
- fast typer, all 10 letters in 2s: 2+0.5s debounce = 2.5s

* fix: LSDV-4600: Lead time for TextArea

* add tests on text area lead_time

* clear helper and remove wait between new tags

* fix lead_time test e2e

* date.now over newDate

* Draft of saving lead_time for per-regions

It stores lead_time in the meta field of the main region
instead of storing it in shared TextArea tag or its ephemeral TextAreaRegions.

* Fix and update ClassificationMixed example

* rename text area regions, remove console

* rename debounce interval

* date.now over newDate

* date.now over newDate in text area

* remove unused toStateJSON

* linting and +Date.now() fixes

* Improve LeadTime mixin

- TypeScript!
- no debounce, it's detected during calculations
- two different approaches on how to count trailing events
- schema with explanations
- simpler usage

* Another try to fix per-region/global textareas

* Firts test for lead_time (and textarea)

* Fix counting time in per-regions

lead_time was not added up because of typo

* Add test for per-regions

* Store all lead_time in results

Information in tags can be easily destroyed,
the only source of truth is results/regions, so store data there.

* TextareaRegionView -> separate file and tidy up

* Fix push

* Fix e2e test by adding lead_time to samples

* Add FF for Lead Time

* Switch to inertial, it feels more natural

* Small fixes

Remove useless metaValue — all data in results and this meta were
overrided by meta from result
Add check for flag enabled

* Fix FF usage

* Fix FF in E2E test for LeadTime

* Move lead_time to results

* No *StateJSON methods!

* Use proper assert in `lead_time` test

* Remove accidential `only` in test

---------

Co-authored-by: Travis1282 <travisjosephclark@gmail.com>
Co-authored-by: hlomzik <hlomzik@gmail.com>
Co-authored-by: hlomzik <hlomzik@users.noreply.github.com>